% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{Build design matrix for B-spline basis.}
\usage{
bspline(x, xl, xr, ndx, bdeg)
}
\arguments{
\item{x}{vector of data.}

\item{xl}{lower bound for knots.}

\item{xr}{upper bound for knots.}

\item{ndx}{number of inner knots.}

\item{bdeg}{degree of B-Spline basis (3 by default).}
}
\value{
B design matrix of B-Spline basis.

knots vector of inner and outer knots.
}
\description{
Build design matrix for B-spline basis.
}
