% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Omega_ar1.R
\name{build_Omega_ar1}
\alias{build_Omega_ar1}
\title{Build Omega AR(1) matrix in spatio-temporal models.}
\usage{
build_Omega_ar1(phi, ntime)
}
\arguments{
\item{phi}{Autorregresive parameter.}

\item{ntime}{Number of temporal periods.}
}
\value{
Omega = Matrix of Var-Cov in time for the noise (without sigma_u^2)

Omegainv = Omega^(-1)

der_sig2u_Omega_phi=\eqn{\frac{\partial \sigma_u^2*Omega}{\partial phi}}

der2_sig2u_Omega_phi=\eqn{\frac{\partial^2 \sigma_u^2*Omega}{\partial phi^2}}
}
\description{
Build Omega AR(1) matrix in spatio-temporal models.
}
