% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.psar.R
\name{coef.psar}
\alias{coef.psar}
\title{Extract method for model coefficients from objects of class psar.}
\usage{
\method{coef}{psar}(object, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{z}{object of class \emph{psar}.}
}
\value{
A vector of the coefficients of the model including fixed
  and random effects.
}
\description{
Extract method for model coefficients from objects of class psar.
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it
 ######################  GAM pure
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20)
gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
summary(gamsar)
coef(gamsar)
}
\seealso{
\itemize{
  \item \code{\link{logLik.psar}} Extract log-likelihood or restricted
    log-likelihood for fitted \emph{psar} objects.
}
}
