% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R, R/effects_nopar.R
\name{eff_nopar}
\alias{eff_nopar}
\alias{eff_nopar}
\title{Compute Total, Direct and Indirect effects functions for  geoadditive
spatial or spatio-temporal semiparametric PS-SAR regression models.}
\usage{
eff_nopar(sptsarfit, variables, conflevel = 0.95)

eff_nopar(sptsarfit, variables, conflevel = 0.95)
}
\arguments{
\item{conflevel}{Numerical value for the confidence interval of the
effect functions.}

\item{fitted_nopar}{A matrix including the non-parametric fitted
functions (GAM) for each covariate.}

\item{sd_fitted_nopar}{A matrix including the standard deviations of
non-parametric functions(GAM) for each convariate.}

\item{Wsp}{A neighbour spatial matrix.
The dimension of the matrix is always \emph{nxn}, where \emph{n} is the
dimension of each spatial coordinate. Default NULL.}

\item{rho}{Spatial parameter of the spatial lag of the dependent
variable for SAR models. Default 0.}

\item{durbin}{A logical value indicating if the model include a spatial lag
of each covariate (either parametric or non-parametric). Default FALSE.}

\item{conf_level}{Numerical value for the confidence interval of the
effect functions.}

\item{fitted_nopar}{A matrix including the non-parametric fitted
functions (GAM) for each covariate.}

\item{sd_fitted_nopar}{A matrix including the standard deviations of
non-parametric functions(GAM) for each convariate.}

\item{Wsp}{A neighbour spatial matrix.
The dimension of the matrix is always \emph{nxn}, where \emph{n} is the
dimension of each spatial coordinate. Default NULL.}

\item{rho}{Spatial parameter of the spatial lag of the dependent
variable for SAR models. Default 0.}

\item{durbin}{A logical value indicating if the model include a spatial lag
of each covariate (either parametric or non-parametric). Default FALSE.}
}
\description{
Compute effects function for non-parametric covariates in semiparametric
models.

Compute effects function for non-parametric covariates in semiparametric
models.
}
