% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_par.R
\name{eff_par}
\alias{eff_par}
\title{Compute direct, indirect and total effects (or impacts)
  for continous parametric covariates in PS-SAR regression models.}
\usage{
eff_par(sptsarfit, variables, nrep = 1000, seed = 1111, m = 100,
  p = 50, tol = 0.01)
}
\arguments{
\item{sptsarfit}{A \emph{psar} object fitted using \code{\link{psar}} function}

\item{variables}{vector including names of non-parametric covariates.}

\item{nrep}{number of repetitions for the simulation. Default 1000.}

\item{seed}{initial seed to get random numbers. Must be set to a 
specific value to make reproducible results. Default 1111.}

\item{m}{number of powers to compute a vector of traces of powers
of a spatial weight matrix (see \code{\link[spdep]{trW}} in 
\pkg{spdep} package). Default 100.}

\item{p}{number of samples used in MC simulation of traces
of a spatial weight matrix (see \code{\link[spdep]{trW}} in
\pkg{spdep} package).Default 50.}

\item{tol}{tolerance (relative to largest variance) for numerical lack 
of positive-definiteness in Sigma when simulate \eqn{\betas}
from the maximum likelihood estimates (see \code{\link[MASS]{mvrnorm}}
in \pkg{MASS} package). Default 0.01.}
}
\value{
An object of class \emph{par.eff.psar}. Can be printed
        with \code{summary}.
        
        The object returned is a list with 3 matrices including
        the results of simulated effects:
         \tabular{ll}{
           \emph{tot_eff} \tab Matrix including simulated total effects 
                               for each variable in rows. \cr
           \emph{dir_eff} \tab Matrix including simulated direct effects 
                               for each variable in rows. \cr
           \emph{ind_eff} \tab Matrix including simulated indirect effects 
                               for each variable in rows. \cr
         }
}
\description{
Compute direct, indirect and total effects (also named
  impacts) for non-parametric covariates included in a semiparametric spatial
  or spatio-temporal SAR model. This model must include a spatial
  lag of the dependent variable (SAR) to have indirect effects 
  different from 0, otherwise, total and direct effects 
  are the same.
}
\details{
DESCRIBE ALGORITHM TO SIMULATE PARAMETRIC EFFECTS
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it

######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
 summary(gamsar)
 ###### Parametric Total, Direct and Indirect Effects
 list_varpar <- c("partrate","agri","cons")
 eff_parvar <- eff_par(gamsar,list_varpar)
 summary(eff_parvar)
 
######################   PSAR-ANOVA with spatial trend
form2 <- unrate ~ partrate + agri + cons +
                  pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                  pspt(long,lat,nknots=c(20,20),psanova=TRUE,
                  nest_sp1=c(1,2),nest_sp2=c(1,2))
##### Spatial trend fixed for period 1996-2014
geospanova_sar <- psar(form2,data=unemp_it,Wsp=Wsp_it,sar=TRUE,
                   control=list(thr=1e-1,maxit=200,trace=FALSE))
summary(geospanova_sar)
 ###### Parametric Total, Direct and Indirect Effects
 list_varpar <- c("partrate","agri","cons")
 eff_parvar <- eff_par(geospanova_sar,list_varpar)
 summary(eff_parvar)
 
######################   PSAR-ANOVA with spatio-temporal trend and 
######################   temporal autorregresive noise
 form3 <- unrate ~ partrate + agri + cons +
                   pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                   pspt(long,lat,year,nknots=c(18,18,8),psanova=TRUE,
                   nest_sp1=c(1,2,3),nest_sp2=c(1,2,3),
                   nest_time=c(1,2,2),ntime=19)
sptanova_sar_ar1 <- psar(form3,data=unemp_it,Wsp=Wsp_it,sar=TRUE,ar1=TRUE,
                    control=list(thr=1e-1,maxit=200,trace=FALSE))
summary(sptanova_sar_ar1)
 ###### Parametric Total, Direct and Indirect Effects
 list_varpar <- c("partrate","agri","cons")
 eff_parvar <- eff_par(sptanova_sar_ar1,list_varpar)
 summary(eff_parvar)

}
\references{
\itemize{ 
  \item LeSage, J. and Pace, K. (2009). \emph{Introduction to 
        Spatial Econometrics}. CRC Press, Boca Raton.
 }
}
\seealso{
\itemize{
  \item \code{\link{psar}} estimate spatial or spatio-temporal 
          semiparametric PS-SAR regression models.
  \item \code{\link{eff_nopar}} compute total, direct and indirect effect
          functions for non-parametric continuous covariates.
  \item \code{\link{fit_terms}} compute smooth functions for non-parametric
          continuous covariates.
  \item \code{\link[spdep]{impacts}} similar function in \pkg{spdep} 
          package to compute impacts in spatial parametric econometric 
          models.                              
}

Other Direct, Indirect and Total Effects.: \code{\link{eff_nopar}},
  \code{\link{plot_eff_nopar}}
}
\concept{Direct, Indirect and Total Effects.}
\keyword{Direct}
\keyword{Indirect}
\keyword{SAR,}
\keyword{covariates.}
\keyword{effects,}
\keyword{parametric}
