% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.psar.R
\name{logLik.psar}
\alias{logLik.psar}
\title{Extract method log-likelihood or restricted log-likelihood
  from objects of class psar.}
\usage{
\method{logLik}{psar}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{object of class \emph{psar}.}

\item{REML}{logical value indicating if the log-likelihood is
restricted. Default: FALSE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \emph{logLik}. Apart from the
  log-likelihood (or restricted log-likelihood) value, this object
  has some additional attributes:
\tabular{ll}{
  \code{nall} \tab Whole number of observations. \cr
  \code{nobs} \tab Effective number of observations. \cr
  \code{df} \tab Effective degrees of freedom. \cr
 }
}
\description{
Extract method log-likelihood or restricted log-likelihood
  from objects of class psar.
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it
 ######################  GAM pure
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20)
gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
summary(gamsar)
llik_gamsar <- logLik(gamsar)
llikreml_gamsar <- logLik(gamsar, REML = TRUE)
cat("Log-Likelihood: ",llik_gamsar,"\\n")
cat("Restricted Log-Likelihood: ",llikreml_gamsar,"\\n")
}
\seealso{
\itemize{
  \item \code{\link{coef.psar}} Extract vector of coefficients for
    fitted \emph{psar} objects including fixed and random effects.
}
}
