% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM_basis.R
\name{MM_basis}
\alias{MM_basis}
\title{Build mixed-model matrices for P-Spline models.}
\usage{
MM_basis(x, xl, xr, ndx, bdeg, pord, decom = 1)
}
\arguments{
\item{x}{vector of data.}

\item{xl}{lower bound for knots.}

\item{xr}{upper bound for knots.}

\item{ndx}{number of inner knots.}

\item{bdeg}{degree of B-Spline basis (3 by default).}

\item{pord}{degree of penalty differences (2 by default).}

\item{decom}{type of fixed effect matrix \code{X}.}

\item{\itemize{}{\item 1 (default) \code{X=P*U} \item 2 \code{X=(1,x,...,x^{pord-1})} \item 3 FILL }}
}
\value{
X design matrix of fixed-effects in mixed models.

Z design Matrix of random-effects in mixed models.

d vector of eigenvalues of SVD decomposition.

B design matrix of B-Spline basis.

m number of columns of B matrix.

D penalty matrix of differences.

U matrix including all vectors of null-space of B.

\code{knots} number of knots.
}
\description{
Build mixed-model matrices for P-Spline models.
}
