% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_terms.R
\name{plot_terms}
\alias{plot_terms}
\title{Plot terms of non-parametric covariates in PS-SAR 
  regression models.}
\usage{
plot_terms(fitterms, data, conflevel = 0.95)
}
\arguments{
\item{fitterms}{object returned from \code{\link{fit_terms}} function.}

\item{data}{dataframe with the data.}

\item{conflevel}{numerical value for the confidence interval of the
term. Default 0.95.}
}
\value{
plot of the terms for each non-parametric covariate included 
 in the object returned from \code{\link{fit_terms}}.
}
\description{
For each non-parametric covariate the plot of the term
  includes confidence intervals and the decomposition in fixed and 
  random part when the term is reparameterized as a mixed model.
}
\examples{
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it
######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
 summary(gamsar)
 ######################  Fit non-parametric terms (spatial trend must be name "spttrend")
 list_varnopar <- c("serv","empgrowth")
 terms_nopar <- fit_terms(gamsar,list_varnopar)
 ######################  Plot non-parametric terms
 plot_terms(terms_nopar,unemp_it)
 
 @export
}
\references{
\itemize{ 
  \item Wood, S.N. (2017). \emph{Generalized Additive Models. 
  An Introduction with \code{R}} (second edition). CRC Press, Boca Raton.
 }
}
\seealso{
\itemize{
  \item \code{\link{fit_terms}} compute smooth functions for non-parametric
                                continuous covariates.
  \item \code{\link{plot_effects_nopar}} plot the effects functions 
    of non-parametric covariates.
  \item \code{\link[mgcv]{plot.gam}} plot the terms fitted by 
    \code{\link[mgcv]{gam}} function in \pkg{mgcv} package.   
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
