% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timetrend.R
\name{plot_timetrend}
\alias{plot_timetrend}
\title{Plot time trend for each spatial unit.}
\usage{
plot_timetrend(sptsarfit, data, time = "year", spat = "region",
  xlab = "Time", ylab = "Time Trend",
  title = "Non-Parametric Time Trend by Spatial Unit")
}
\arguments{
\item{sptsarfit}{\emph{psar} object fitted using \code{\link{psar}} function.}

\item{data}{dataframe with the data.}

\item{time}{name of variable including the temporal coordinate. Default "year".}

\item{spat}{name of variable including the spatial units. Default "region".}

\item{xlab}{label of x-axis. Default "Time".}

\item{ylab}{label of y-axis. Default "Time Trend".}

\item{title}{title of graphics. Default "Non-Parametric Time Trend by Spatial Unit"}
}
\description{
In models including a spatio-temporal trend plot a time
  trend for each spatial unit in the dataframe. 
  It can be useful to detect heterogeneous pattern of time trends for each
  spatial unit.
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it
 ###############################################
 # Spatio-temporal semiparametric ANOVA model without spatial lag
 # Interaction terms f12,f1t,f2t and f12t with nested basis
 # Remark: It is necessary to include ntime as argument
 # Remark: nest_sp1, nest_sp2 and nest_time must be divisors of nknots
 form4 <- unrate ~ partrate + agri + cons +
                   pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                   pspt(long,lat,year,nknots=c(18,18,8),psanova=TRUE,
                   nest_sp1=c(1,2,3),nest_sp2=c(1,2,3),
                   nest_time=c(1,2,2),ntime=19)
 sptanova <- psar(form4,data=unemp_it,
                  control=list(thr=1e-2,maxit=200,trace=FALSE)) 
 plot_timetrend(sptanova,data=unemp_it,time="year",
               spat="name")
               
}
\seealso{
\itemize{
  \item \code{\link{plot_main_spt}} plot main terms of non-parametric 
    spatial (2d) or spatio-temporal (3d) trends in ANOVA models.
  \item \code{\link[mgcv]{plot.gam}} plot the terms fitted by 
    \code{\link[mgcv]{gam}} function in \pkg{mgcv} package.   
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
