% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.par.eff.psar.R,
%   R/print.summary.psar.R
\name{print.summary.psar}
\alias{print.summary.psar}
\alias{print.summary.par.eff.psar}
\title{Print method for objects of class summary.par.eff.psar.}
\usage{
\method{print}{summary.par.eff.psar}(x, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{print}{summary.psar}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.par.eff.psar}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}

\item{x}{object of class \emph{summary.psar}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for objects of class summary.par.eff.psar.

Print method for objects of class summary.psar.
}
\examples{
  See examples for \\code{\\link{effects_par}} function.
  See examples for \\code{\\link{psar}} function.
}
\seealso{
\itemize{
  \item \code{\link{eff_par}} Compute direct, indirect and
    total effects (or impacts).
    for continous parametric covariates in PS-SAR regression models.
  \item \code{\link{summary.par.eff.psar}} Summary method
    for \emph{par.eff.psar} objects.
}

\itemize{
  \item \code{\link{summary.psar}} Summary method for \emph{psar} objects.
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}

Roman Minguez \email{roman.minguez@uclm.es}
}
