% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.par.eff.psar.R
\name{summary.par.eff.psar}
\alias{summary.par.eff.psar}
\title{Summary method for object of class par.eff.psar.}
\usage{
\method{summary}{par.eff.psar}(object, ...)
}
\arguments{
\item{object}{\emph{par.eff.psar} object fitted using \code{\link{psar}} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \emph{summary.par.eff.psar}
}
\description{
This method summarizes direct, indirect and total effects (or impacts)
  for continous parametric covariates in PS-SAR regression models.
}
\examples{
  See examples for \\code{\\link{effects_par}} function.
}
\seealso{
\itemize{
  \item \code{\link{eff_par}} Compute direct, indirect and total
    effects (or impacts)
    for continous parametric covariates in PS-SAR regression models.
  \item \code{\link{print.summary.par.eff.psar}} print objects of
    class \emph{summary.par.eff.psar}
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
