% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_nopar.R, R/fit_terms.R
\name{eff_nopar}
\alias{eff_nopar}
\alias{eff_nopar}
\alias{fit_terms}
\title{Compute direct, indirect and total effect functions 
  for continous non-parametric covariates in spatial or 
  spatio-temporal semiparametric PS-SAR regression models.}
\usage{
eff_nopar(sptsarfit, variables, conflevel = 0.95)

fit_terms(sptsarfit, variables)
}
\arguments{
\item{sptsarfit}{\emph{psar} object fitted using \code{\link{psar}} function}

\item{variables}{vector including names of non-parametric covariates.}

\item{conflevel}{mumerical value for the confidence interval of the
effect functions. Default 0.95.}

\item{sptsarfit}{\emph{psar} object fitted using \code{\link{psar}} function}

\item{variables}{vector including names of non-parametric covariates.}
}
\value{
A list including
  \tabular{ll}{
    \emph{effnopar_tot} \tab Matrix including total effects in columns. \cr
    \emph{effnopar_dir} \tab Matrix including direct effects in columns. \cr
    \emph{effnopar_ind} \tab Matrix including indirect effects in columns. \cr
    \emph{effnopar_tot_up} \tab Matrix including upper bounds of total effects in columns. \cr
    \emph{effnopar_dir_up} \tab Matrix including upper bounds of direct effects in columns. \cr
    \emph{effnopar_ind_up} \tab Matrix including upper bounds of indirect effects in columns. \cr
    \emph{effnopar_tot_low} \tab Matrix including lower bounds of total effects in columns. \cr
    \emph{effnopar_dir_low} \tab Matrix including lower bounds of direct effects in columns. \cr
    \emph{effnopar_ind_low} \tab Matrix including lower bounds of indirect effects in columns. \cr
 }

CONTINUAR AQUÍ
}
\description{
Compute direct, indirect and total effect functions for 
  non-parametric covariates included in a semiparametric spatial
  or spatio-temporal SAR model. This model must include a spatial
  lag of the dependent variable (SAR) to have indirect effects 
  different from 0, otherwise, total and direct function effects 
  are the same.

CONTINUAR AQUÍ
}
\details{
DESCRIBE ALGORITHM TO COMPUTE EFFECT FUNCTIONS AND THE 
         SMOOTHING TO PLOT
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it

######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
 summary(gamsar)
 ###### Non-Parametric Total, Direct and Indirect Effects
 list_varnopar <- c("serv","empgrowth")
 eff_nparvar <- eff_nopar(gamsar,list_varnopar)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=TRUE)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=FALSE)
 
######################   PSAR-ANOVA with spatial trend
form2 <- unrate ~ partrate + agri + cons +
                  pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                  pspt(long,lat,nknots=c(20,20),psanova=TRUE,
                  nest_sp1=c(1,2),nest_sp2=c(1,2))
##### Spatial trend fixed for period 1996-2014
geospanova_sar <- psar(form2,data=unemp_it,Wsp=Wsp_it,sar=TRUE,
                   control=list(thr=1e-1,maxit=200,trace=FALSE))
summary(geospanova_sar)
 ###### Non-Parametric Total, Direct and Indirect Effects
 list_varnopar <- c("serv","empgrowth")
 eff_nparvar <- eff_nopar(geospanova_sar,list_varnopar)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=TRUE)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=FALSE)
 
######################   PSAR-ANOVA with spatio-temporal trend and 
######################   temporal autorregresive noise
 form3 <- unrate ~ partrate + agri + cons +
                   pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                   pspt(long,lat,year,nknots=c(18,18,8),psanova=TRUE,
                   nest_sp1=c(1,2,3),nest_sp2=c(1,2,3),
                   nest_time=c(1,2,2),ntime=19)
sptanova_sar_ar1 <- psar(form3,data=unemp_it,Wsp=Wsp_it,sar=TRUE,ar1=TRUE,
                    control=list(thr=1e-1,maxit=200,trace=FALSE))
summary(sptanova_sar_ar1)
 ###### Non-Parametric Total, Direct and Indirect Effects
 list_varnopar <- c("serv","empgrowth")
 eff_nparvar <- eff_nopar(sptanova_sar_ar1,list_varnopar)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=TRUE)
 plot_effects_nopar(eff_nparvar,unemp_it,smooth=FALSE)

}
\references{
\itemize{ 
  \item Basile, R., Durbán, M., Mínguez, R., Montero, J.
        M., and Mur, J. (2014). Modeling regional economic 
        dynamics: Spatial dependence, spatial heterogeneity and 
        nonlinearities. \emph{Journal of Economic Dynamics and 
        Control}, (48), 229-245.
        
  \item LeSage, J. and Pace, K. (2009). \emph{Introduction to 
        Spatial Econometrics}. CRC Press, Boca Raton.
        
  \item Mínguez, R.; Basile, R. and Durbán, M. (2018). An Alternative Semiparametric Model
        for Spatial Panel Data. Under evaluation in \emph{Statistical
        Methods and Applications}.
 }

\itemize{ 
  \item Wood, S.N. (2017). \emph{Generalized Additive Models. 
  An Introduction with \code{R}} (second edition). CRC Press, Boca Raton.
 }       
#' @examples
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(sptpsar)
data(unemp_it); Wsp <- Wsp_it
######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- psar(form1,data=unemp_it,sar=TRUE,Wsp=Wsp_it)
 summary(gamsar)
 ######################  Fit non-parametric terms (spatial trend must be name "spttrend")
 list_varnopar <- c("serv","empgrowth")
 terms_nopar <- fit_terms(gamsar,list_varnopar)
 ######################  Plot non-parametric terms
 plot_terms(terms_nopar,unemp_it)
 
 @export
}
\seealso{
\itemize{
  \item \code{\link{psar}} estimate spatial or spatio-temporal semiparametric PS-SAR
  regression models.
  \item \code{\link{eff_par}} compute and simulate total, direct and indirect effect
                               (or impacts) for parametric continuous covariates.
  \item \code{\link{fit_terms}} compute smooth functions for non-parametric
                                continuous covariates.
  \item \code{\link{plot_effects_nopar}} plot the non-parametric effects functions
                                         allowing for previous smoothing.
}

\itemize{
  \item \code{\link{psar}} estimate spatial or spatio-temporal semiparametric PS-SAR
          regression models.
  \item \code{\link{eff_nopar}} compute total, direct and indirect effect
          functions for non-parametric continuous covariates.
  \item \code{\link{plot_terms}} plot the fitted terms.
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}

Roman Minguez \email{roman.minguez@uclm.es}
}
\keyword{Direct}
\keyword{Indirect}
\keyword{SAR,}
\keyword{covariates.}
\keyword{effects,}
\keyword{non-parametric}
